requirements.txt
Django==3.1.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djangorestframework-jwt ==1.10.0
python-dotenv
gunicorn==20.0.4
django-heroku==0.3.1
django-cors-headers==3.7.0